Fix plugin position parameter to have a lower bound of 0
the old code starts position parameters from 1, the new from 0. When
reading from a preset file the position value of the first plugin was
changed to 1 by the json file reader, resulting in the first 2 plugins
having the same position value and a more or less random ordering of
those plugins.