replaced lfo in vibe.cc with faust based lfo and changed license
scaled univibe frequency control by factor of 1/60.

Copy of email message with license change permission:

Message-ID: <4F6C8B56.9070407@gmail.com>
Date: Fri, 23 Mar 2012 07:40:22 -0700
From: Ryan Billing <ryjobil atdomain gmail dt com>
To: Andreas Degert <andreas.degert@googlemail.com>,
 Josep Andreu <holborn domain telefonica dt net>
Subject: Re: Vibe

As far as the Vibe goes, I am the sole author of it, so I have no
problem letting you "rebrand" it to GPLV2+ any later version of GPL.

I don't think this licensing is intentionally stuck on the GPL2, but I
am forwarding Holborn for history there.  Holborn was involved from the
beginning of the project and he will know why the licensing was chosen
as it is.

If Holborn is amiable (if he does not have specific reasons to keep the
current licensing), then perhaps we can update Rakarrack to GPLV2 and
any later version of GPL.  I personally trust any version of the GPL to
keep the code free, and keep me safe from being sued for consequences
due to its use.

So, this is my written permission to modify the text on the headers of
your copies of the Vibe code to indicate any GPL revision.  If you need
me to specifically modify and issue something to you, then that also is
fine.

I hope that helps.
Best Regards,
Ryan