wscript: add option --dist-tree
"./waf dist" creates an archive "guitarix2-XXXX.tar.bz2" for a git
tree-ish.

When no option --dist-tree is given, the tree-ish is the tag created
from VERSION (in wscript) by prepending a V. XXXX is VERSION

When a version tag is given as value for --dist-tree (name starting
with V and a digit as second character) , it's used to created the
archive, XXXX is tag with the leading V.

When another value is give for --dist-tree, it is used as tree-ish for
git archive, XXXX is git- plus the commit hash truncated to 8 chars.

Examples:

./waf dist   ->   error because tag V0.22beta2-git not found
                  (use this to create a release file directly
		   after setting a tag)

./waf dist --dist-tree V0.22beta2  -> guitarix2-V0.22beta2.tar.bz2

./waf dist --dist-tree HEAD        -> guitarix2-git-c6b044f3.tar.bz2