replace ugly offscreen window bug workaround with a simple one
The problem seems to be in libcanberra-gtk-module, which installs
idle-handlers and tries to get an xid while not noticing that the
GtkWindow is really a GtkOffscreenWindow. The problem only surfaces
when the canberra module is loaded, an event loop is executed while
the offscreen window exists and also seems to depend on the used X
server and gtk/gdk library versions (either just gtk warnings, or
nothing, or crash...).

The problem can be circumvented by setting a type hint that makes the
window uninteresting for libcanberra-gtk-module.