small change to client parameter change signalling
in debug build: when env variable  GUITARIX_TRACE is set (no value,
just existence), a trace of all parameter changes is written to
stderr. "blocked" means the change is not sent to the server (because
it was originated by the server). If the variable has a value it is
interpretet as the id of the parameter to trace.