client/server: convolver
Convolvers should work now in client/server mode. Since the file
selection window can only display local files (or data on real remote
file systems supported by a virtual filesystem backend), the file
selector window is not supported in the client mode. Instead a list
displaying the subdirectory tree of the system and user IR dirs is
shown. This list is not yet synchronized with the currently selection
directory.