store IR path info in a system independent way
The old scheme auf leaving the directory variable empty doesn't work
with the new subdirectories. Now the path prefix to the Guitarix IR
installation directory is substituted with %S and path prefix to the
IR dir in the user home dir with %U when storing the data. On reload
the symbols are substituted with the correct path prefixes for the
current system. For old presets the legacy method is still supported.

New bugs in ladspa_guitarix fixed, some things re-organized
(BasicOptions as base class for CmdlineOptions introduced).