extend LADSPA loader to LV2 plugins
adds a dependency on the library lilv and extends the existing LADSPA
loaded so that it can also handle LV2 plugins. Only plugins with 1/1
or 2/2 audio channels will be displayed. Plugins with non-optional
connections other than audio and control will also be excluded (e.g.
with required MIDI event connection).

More testing and some polishing needed. Non-optional feature requests
of plugins are ignored in this release.