waf2: fixed build for configure --convolver--ffmpeg
convolver and resampler can now be added to the link phase by just
adding the symbols ZITA_CONVOLVER or ZITA_RESAMPLER to the "use" list
of the task generator. The configuration options define if the symbol
adds a dependency to an external library or adds a list of object
files. The use-system takes also care of recursively adding libraries.