wscript: set check_in_faust_version to 2.20.2
The purpose of the variable check_in_faust_version is to document
which Faust version is used to generate the checked in files.

The variable other_faust_versions contains Faust versions that should
also work (at the moment 2.15.11), but those versions should not be
used to generate files for check in.

Also added --no-version-header to the invocation of dsp2cc, so there
is no mass change of generated files when the Faust version changes
but still produces the same output. Faust 2.20 produce quite stable
output.